Break exception macros into a file of their own.
