+++ include/deal.II/sundials/arkode.h
@@ -56,9 +56,6 @@
 DEAL_II_NAMESPACE_OPEN
 
 
-// Shorthand notation for ARKODE error codes.
-#  define AssertARKode(code) Assert(code >= 0, ExcARKodeError(code))
-
 /**
  * A namespace for dealing with ODE solvers through the SUNDIALS package.
  */
